+++ packages/react/src/popover/popover.css.ts
@@ -8,12 +8,11 @@ const native = recipe({
     boxSizing: "border-box",
     margin: 0,
     padding: 0,
-
-
     width: "100vw",
     height: "100vh",
     maxWidth: "100%",
     maxHeight: "100%",
+    overflow: "hidden",
   },
   variants: {
     open: {
@@ -30,8 +29,10 @@ const native = recipe({
 });
 const portal = recipe({
   base: {
+    boxSizing: "border-box",
     position: "fixed",
     inset: 0,
+    overflow: "hidden",
   },
   variants: {
     open: {