+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @star4/react
 
+## 0.1.6
+
+### Patch Changes
+
+- 5c684eb: popover: hide overflow
+
+## 0.1.5
+
+### Patch Changes
+
+- 3a9fd9c: action-button: add forwarded attributes
+
 ## 0.1.4
 
 ### Patch Changes